Refactor chat view
