+++ commet/integration_test/matrix/change_space_name_test.dart
@@ -1,7 +1,7 @@
 import 'package:commet/ui/atoms/space_header.dart';
 import 'package:commet/ui/atoms/space_icon.dart';
 import 'package:commet/ui/organisms/space_summary/space_summary_view.dart';
-import 'package:commet/ui/pages/chat/chat_page.dart';
+import 'package:commet/ui/pages/main/main_page.dart';
 import 'package:commet/ui/pages/settings/desktop_settings_page.dart';
 import 'package:commet/utils/rng.dart';
 import 'package:flutter/material.dart';
@@ -23,10 +23,10 @@ void main() {
     await tester.login(app);
 
     await _selectSpace(tester);
-    ChatPageState chatPage = tester.state(find.byType(ChatPage));
+    MainPageState chatPage = tester.state(find.byType(MainPageState));
 
     String newName = "New Space Name ${RandomUtils.getRandomString(10)}";
-    var space = chatPage.selectedSpace!;
+    var space = chatPage.currentSpace!;
     String name = space.displayName;
 
     expect(name, isNot(newName));
