+++ commet/lib/ui/molecules/message_input.dart
@@ -5,7 +5,7 @@ import 'package:commet/config/build_config.dart';
 import 'package:commet/main.dart';
 import 'package:commet/ui/molecules/attachment_icon.dart';
 import 'package:commet/ui/molecules/emoticon_picker.dart';
-import 'package:commet/ui/pages/chat/chat_page.dart';
+import 'package:commet/ui/organisms/chat/chat.dart';
 import 'package:commet/client/components/emoticon/emoji_pack.dart';
 import 'package:commet/client/components/gif/gif_search_result.dart';
 import 'package:file_picker/file_picker.dart';
